requirements.txt
django==3.1
psycopg2-binary==2.8.5
pillow==7.1
django-shopping-cart==0.1
